Silvio ajuste componente

+++ app/src/main/java/com/example/tophair/app/data/entities/ServicoEntity.kt
@@ -5,6 +5,6 @@ data class Servico(
     val nomeServico: String? = null,
     val categoria: String? = null,
     val descricao: String? = null,
-    val preco: Int? = null,
+    val preco: Double? = null,
     val qtdTempoServico: String? = null
 )
